+++ shuup/admin/modules/orders/views/payment.py
@@ -10,7 +10,7 @@ from __future__ import unicode_literals
 from django import forms
 from django.contrib import messages
 from django.http.response import HttpResponseRedirect
-from django.utils.translation import ugettext as _
+from django.utils.translation import ugettext_lazy as _
 from django.views.generic import DetailView, UpdateView
 from django.views.generic.edit import BaseDeleteView