+++ shuup/admin/modules/products/views/toolbars.py
@@ -7,7 +7,7 @@
 # LICENSE file in the root directory of this source tree.
 from __future__ import unicode_literals
 
-from django.utils.translation import ugettext as _
+from django.utils.translation import ugettext_lazy as _
 from enumfields import Enum
 
 from shuup.admin.toolbar import (