+++ shuup/admin/toolbar.py
@@ -12,7 +12,7 @@ import six
 from django.middleware.csrf import get_token
 from django.utils.html import conditional_escape, format_html
 from django.utils.safestring import mark_safe
-from django.utils.translation import ugettext as _
+from django.utils.translation import ugettext_lazy as _
 
 from shuup.admin.utils.forms import flatatt_filter
 from shuup.admin.utils.permissions import get_missing_permissions