+++ shuup/admin/modules/orders/sections.py
@@ -8,7 +8,7 @@
 from __future__ import unicode_literals
 
 from django.conf import settings
-from django.utils.translation import ugettext as _
+from django.utils.translation import ugettext_lazy as _
 
 from shuup.admin.base import Section
 from shuup.admin.utils.permissions import get_missing_permissions