+++ shuup/admin/modules/orders/views/refund.py
@@ -5,13 +5,11 @@
 #
 # This source code is licensed under the OSL-3.0 license found in the
 # LICENSE file in the root directory of this source tree.
-from __future__ import unicode_literals
-
 from django import forms
 from django.conf import settings
 from django.contrib import messages
 from django.http.response import HttpResponseRedirect
-from django.utils.translation import ugettext as _
+from django.utils.translation import ugettext_lazy as _
 from django.views.generic import UpdateView
 
 from shuup.admin.supplier_provider import get_supplier