+++ shuup/front/utils/companies.py
@@ -8,7 +8,7 @@
 from __future__ import unicode_literals
 
 from django.core.exceptions import ValidationError
-from django.utils.translation import ugettext as _
+from django.utils.translation import ugettext_lazy as _
 
 from shuup import configuration
 from shuup.core.utils import tax_numbers