+++ shuup/notify/actions/email.py
@@ -11,7 +11,7 @@ import logging
 from django import forms
 from django.conf import settings
 from django.core.mail.message import EmailMessage
-from django.utils.translation import ugettext as _
+from django.utils.translation import ugettext_lazy as _
 from html import unescape
 
 from shuup.admin.forms.widgets import CodeEditorWithHTMLPreview